actually no suffix option

+++ Program.cs
@@ -2,6 +2,8 @@
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using System.Xml.Linq;
+using static System.Net.Mime.MediaTypeNames;
 
 namespace RiffNamer
 {
@@ -56,6 +58,8 @@ namespace RiffNamer
                 //Filename for logs
                 string fileName = Path.GetFileName(arg);
                 string fileNameNoExt = Path.GetFileNameWithoutExtension(arg);
+                string dir = Path.GetDirectoryName(arg);
+                string wemExt = Path.GetExtension(arg);
 
                 //Extension check
                 /*if (!Path.HasExtension(arg))
@@ -122,7 +126,7 @@ namespace RiffNamer
                                 string tag = new string(reader.ReadChars(4)); //adtl
                                 uint tagSize = reader.ReadUInt32(); //labl
                                 uint unknown = reader.ReadUInt32(); // 1 usually
-                                embeddedFilenameMarker = reader.ReadCString();
+                                embeddedFilenameMarker = reader.ReadCString().Normalize();
                                 reader.AlignStream(2);
                                 break;
                             //Default generic chunk to skip
@@ -152,31 +156,78 @@ namespace RiffNamer
                 };
 
                 DumpEmbeddedFilenameMarkers(embeddedFilenameMarker);
-                if (noSuffix)
-                    Rename(arg, embeddedFilenameMarker);
-                else
-                    PrefixName(arg, embeddedFilenameMarker);
+
+                string newPath = "";
+
+                string newPath_embeddedFileName = dir + "\\" + embeddedFilenameMarker;
+                string newName_embeddedFileNamePrefixed = embeddedFilenameMarker + "_" + fileNameNoExt;
+                string newPath_embeddedFileNamePrefixed = dir + "\\" + newName_embeddedFileNamePrefixed;
+
+                if (!arg.Contains(embeddedFilenameMarker) && !fileNameNoExt.Equals(embeddedFilenameMarker))
+                {
+                    if (noSuffix)
+                    {
+                        if (!File.Exists(newPath_embeddedFileName + wemExt))
+                        {
+                            newPath = newPath_embeddedFileName;
+                        }
+                        else
+                        {
+                            if (!File.Exists(newPath_embeddedFileNamePrefixed + wemExt))
+                            {
+                                newPath = newPath_embeddedFileNamePrefixed;
+                            }
+                        }
+                    }
+                    else
+                    {
+                        if (!File.Exists(newPath_embeddedFileNamePrefixed + wemExt))
+                        {
+                            newPath = newPath_embeddedFileNamePrefixed;
+                        }
+                    }
+                }
+                else if (arg.Contains(embeddedFilenameMarker + "_"))
+                {
+                    string revertedFileName = Path.GetFileNameWithoutExtension(arg).Replace(embeddedFilenameMarker + "_", string.Empty).Normalize();
+                    string newPath_revertedFileName = dir + "\\" + revertedFileName;
+                    if (!File.Exists(newPath_revertedFileName + wemExt))
+                    {
+                        newPath = newPath_revertedFileName;
+                    }
+                }
+                else if (arg.Equals(newPath_embeddedFileName + wemExt))
+                {
+                    Console.WriteLine($"{embeddedFilenameMarker} contains no sound file id to revert to!");
+                    continue;
+                }
+
+                if (newPath=="")
+                {
+                    continue;
+                }
+
+                if (!File.Exists(newPath + wemExt))
+                {
+                    File.Move(arg, newPath + wemExt);
+                    Console.WriteLine($"{arg} renamed to {newPath + wemExt}");
+                }
 
                 foreach (string ext in extraFileExtensons)
                 {
-                    string dir = Path.GetDirectoryName(arg);
-                    string potentialExtraFileNamePath = dir + "\\" + fileNameNoExt + "." + ext;
-                    string potentialExtraFileNamePath2 = dir + "\\" + embeddedFilenameMarker + "." + ext;
-                    if (File.Exists(potentialExtraFileNamePath))
-                        PrefixName(potentialExtraFileNamePath, embeddedFilenameMarker);
-
-                    if (File.Exists(potentialExtraFileNamePath2))
+                    string dotExt = "." + ext;
+                    string extPath = dir + "\\" + fileNameNoExt + dotExt;
+                    if (File.Exists(extPath))
                     {
-                        if (uint.TryParse(fileNameNoExt, out uint soundFileId))
-                            if (noSuffix)
-                                Rename(potentialExtraFileNamePath2, embeddedFilenameMarker);
-                            else
-                                Rename(potentialExtraFileNamePath2, embeddedFilenameMarker + "_" + soundFileId.ToString());
-                        else
-                            Rename(potentialExtraFileNamePath2, fileNameNoExt);
+                        if (!File.Exists(newPath + dotExt))
+                        {
+                            File.Move(extPath, newPath + dotExt);
+                            Console.WriteLine($"{extPath} renamed to {newPath + dotExt}");
+                        }
                     }
                 };
             };
+            Console.Read();
         }
         public static void PrefixName(string path, string name)
         {
@@ -188,7 +239,7 @@ namespace RiffNamer
             {
                 //Rename to originalstringname_numbername
                 File.Move(path, newPath);
-                Console.WriteLine($"Renamed {Path.GetFileName(path)} to {name + "_" + Path.GetFileNameWithoutExtension(path)}");
+                Console.WriteLine($"Prefixed1 {Path.GetFileName(path)} to {name + "_" + Path.GetFileNameWithoutExtension(path)}");
             }
             else
             {
@@ -196,7 +247,7 @@ namespace RiffNamer
                 name = Path.GetFileNameWithoutExtension(path).Replace(name + "_", string.Empty);
                 newPath = dir + "\\" + name + "." + ext;
                 File.Move(path, newPath);
-                Console.WriteLine($"Renamed {Path.GetFileName(path)} to {name}");
+                Console.WriteLine($"Prefixed2 {Path.GetFileName(path)} to {name}");
             }
         }
         public static void Rename(string path, string name)
@@ -205,8 +256,17 @@ namespace RiffNamer
             name = name.Normalize();
             string ext = Path.GetExtension(path).Substring(1);
             string newPath = dir + "\\" + name + "." + ext;
-            File.Move(path, newPath);
-            Console.WriteLine($"Renamed {Path.GetFileName(path)} to {Path.GetFileNameWithoutExtension(newPath)}");
+            if (!File.Exists(newPath))
+            {
+                File.Move(path, newPath);
+                Console.WriteLine($"Renamed {Path.GetFileName(path)} to {Path.GetFileNameWithoutExtension(newPath)}");
+            }
+            else
+            {
+                Console.WriteLine($"{Path.GetFileNameWithoutExtension(newPath)} already exists, prefixing");
+
+                PrefixName(path, name);
+            }
         }
         public static void DumpEmbeddedFilenameMarkers(string name)
         {
